# artifacts/api-server/openapi.yaml
openapi: 3.0.3
info:
  title: ISM Finance API
  version: 1.2.0
  description: API REST do SaaS financeiro com painel JWT e integração Power BI (v1 token).
servers:
  - url: http://localhost:3000/api
    description: Local
...
paths:
...
  /v1/bancos:
    get:
      tags: [PowerBI v1]
      operationId: getV1Bancos
      security: [{ bearerV1Token: [] }]
      responses:
        "200":
          description: Contas bancárias (extração)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
  /v1/contasPagar:
    get:
      tags: [PowerBI v1]
      operationId: getV1ContasPagar
      security: [{ bearerV1Token: [] }]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 10000 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Lançamentos CP com joins para BI
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
  /v1/contasReceber:
    get:
      tags: [PowerBI v1]
      operationId: getV1ContasReceber
      security: [{ bearerV1Token: [] }]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 10000 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Lançamentos CR com joins para BI
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
  /v1/pessoas:
    get:
      tags: [PowerBI v1]
      operationId: getV1Pessoas
      security: [{ bearerV1Token: [] }]
      responses:
        "200":
          description: Parceiros/pessoas
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
  /v1/filiais:
    get:
      tags: [PowerBI v1]
      operationId: getV1Filiais
      security: [{ bearerV1Token: [] }]
      responses:
        "200":
          description: Filiais
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
  /v1/planoContas:
    get:
      tags: [PowerBI v1]
      operationId: getV1PlanoContas
      security: [{ bearerV1Token: [] }]
      responses:
        "200":
          description: Plano de contas
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
  /v1/categoriaPlanoConta:
    get:
      tags: [PowerBI v1]
      operationId: getV1CategoriaPlanoConta
      security: [{ bearerV1Token: [] }]
      responses:
        "200":
          description: Categorias do plano de contas
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    bearerV1Token:
      type: http
      scheme: bearer
      description: Token estático validado na tabela `tokens_api`.
...
  schemas:
    ApiErrorItem:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
    ApiEnvelopeGeneric:
      type: object
      properties:
        data: {}
        meta:
          nullable: true
          type: object
          additionalProperties: true
        errors:
          nullable: true
          type: array
          items:
            $ref: "#/components/schemas/ApiErrorItem"
...